+++ src/main/kotlin/app/revanced/patches/youtube/misc/litho/filter/fingerprints/ComponentContextParserFingerprint.kt
@@ -6,8 +6,8 @@ import com.android.tools.smali.dexlib2.Opcode
 object ComponentContextParserFingerprint : MethodFingerprint(
     opcodes = listOf(
         Opcode.INVOKE_VIRTUAL,
+        Opcode.INVOKE_VIRTUAL,
         Opcode.IPUT_OBJECT,
-        Opcode.NEW_INSTANCE
     ),
     strings = listOf("Component was not found %s because it was removed due to duplicate converter bindings.")
 )